requirements.txt
flask
requests
python-dotenv
langchain
langchain-community
langchain-groq
openai
markdown2